requirements.txt
sounddevice
scipy
SpeechRecognition
requests
python-dotenv
edge-tts
